api/requirements.txt
django==2.1.4
flake8==3.6.0
psycopg2-binary==2.7.6.1
django-extensions==2.1.4
ipython==7.2.0
djangorestframework==3.9.0
